refactor(config): modularize config discovery
